+++ cclient/api/extended/send_trytes.c
@@ -12,7 +12,7 @@
 #include "cclient/api/extended/store_and_broadcast.h"
 
 retcode_t iota_client_send_trytes(iota_client_service_t const* const serv, hash8019_array_p const trytes,
-                                  uint32_t const depth, uint32_t const mwm, flex_trit_t const* const reference,
+                                  uint32_t const depth, uint8_t const mwm, flex_trit_t const* const reference,
                                   bool const local_pow, transaction_array_t* out_transactions) {
   retcode_t ret_code = RC_ERROR;
   get_transactions_to_approve_req_t* tx_approve_req = get_transactions_to_approve_req_new();
