+++ utils/signed_files.h
@@ -18,7 +18,7 @@ extern "C" {
 #endif
 
 retcode_t iota_file_signature_validate(char const *const filename, char const *const signature_filename,
-                                       flex_trit_t const *const public_key, size_t depth, size_t index,
+                                       flex_trit_t const *const public_key, uint32_t depth, size_t index,
                                        bool *const valid);
 
 #ifdef __cplusplus
